extract request to a common util
