+++ packages/dd-trace/src/ci-visibility/exporters/agentless/request.js
@@ -3,49 +3,20 @@
 const https = require('https')
 const http = require('http')
 const log = require('../../../log')
-const { storage } = require('../../../../../datadog-core')
 
-function retriableRequest (data, options, callback) {
-  const store = storage.getStore()
-
-  storage.enterWith({ noop: true })
-
-  const client = options.protocol === 'https:' ? https : http
-
-  const timeout = options.timeout || 15000
-
-  const request = client.request(options, res => {
-    let responseData = ''
-
-    res.setTimeout(timeout)
-
-    res.on('data', chunk => { responseData += chunk })
-    res.on('end', () => {
-      if (res.statusCode >= 200 && res.statusCode <= 299) {
-        callback(null, responseData, res.statusCode)
-      } else {
-        const error = new Error(`Error from the intake: ${res.statusCode} ${http.STATUS_CODES[res.statusCode]}`)
-        error.status = res.statusCode
-
-        callback(error, null, res.statusCode)
-      }
-    })
-  })
-  request.setTimeout(timeout, request.abort)
-  request.write(data)
-
-  storage.enterWith(store)
-
-  return request
-}
+const retriableRequest = require('../../../exporters/common/request')
 
 function request (data, options, callback) {
-  const firstRequest = retriableRequest(data, options, callback)
+  const client = options.protocol === 'https:' ? https : http
+
+  const firstRequest = retriableRequest(options, client, callback)
+  firstRequest.write(data)
 
   // The first request will be retried
   const firstRequestErrorHandler = () => {
     log.debug('Retrying request to the intake')
-    const retriedReq = retriableRequest(data, options, callback)
+    const retriedReq = retriableRequest(options, client, callback)
+    retriedReq.write(data)
     // The retried request will fail normally
     retriedReq.on('error', e => callback(new Error(`Network error trying to reach the intake: ${e.message}`)))
     retriedReq.end()
